# backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: SmartPresence API
  description: API for managing users, rooms, classes, sessions, and attendance in an educational institution
  version: 1.0.0
  contact:
    name: SmartPresence Development Team
    email: dev@smartpresence.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.smartpresence.com/api
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /mobile/me/classes:
    get:
      tags:
        - Mobile
      summary: Get my classes
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Classes list
...
  /classes:
    get:
      tags:
        - Classes
      summary: Get all classes
      description: Retrieve all classes
      responses:
        '200':
          description: Classes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassesResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Classes
      summary: Create class
      description: Create a new class
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClassRequest'
      responses:
        '201':
          description: Class created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClassResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      required:
        - success
        - message
        - timestamp
      properties:
        success:
          type: boolean
        message:
          type: string
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
        - path
        - method
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - message
            - type
            - statusCode
          properties:
            message:
              type: string
            type:
              type: string
            statusCode:
              type: integer
            details:
              type: array
              items:
                type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
        method:
          type: string
        requestId:
          type: string

    # User Schemas
...
    Class:
      type: object
      required:
        - class_id
        - name
        - course_code
        - teacher_id
        - created_at
      properties:
        class_id:
          type: integer
        name:
          type: string
        course_code:
          type: string
        description:
          type: string
          nullable: true
        teacher_id:
          type: integer
        created_at:
          type: string
          format: date-time
    ClassResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Class'
    ClassesResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Class'
    CreateClassRequest:
      type: object
      required:
        - name
        - course_code
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 150
        course_code:
          type: string
          minLength: 1
          maxLength: 50
        description:
          type: string

    # Session Schemas
...